Added translation to saving over the movie file being played. Removed some int/unsigned warnings from mrvPreferences.cpp.
Former-commit-id: b7dbf9737482d0c53f3eb731dbfa4ee8390901b3 [formerly 66ee7d770ee74380318123eb24d7f308503856b8]
Former-commit-id: 935071803f37f76b9967671df9b8a77345a9a2b5

+++ mrViewer/src/gui/mrvPreferences.cpp
@@ -762,7 +762,7 @@ fltk::StyleSet*     newscheme = NULL;
 
     fltk::Preferences subtitles( base, "subtitles" );
     subtitles.get( "font", tmpS, "Arial", 2048 );
-    for (unsigned i = 0; i < uiPrefs->uiPrefsSubtitleFont->children(); ++i )
+    for (int i = 0; i < uiPrefs->uiPrefsSubtitleFont->children(); ++i )
     {
         if ( strcmp( uiPrefs->uiPrefsSubtitleFont->child(i)->label(),
                      tmpS ) == 0 )
@@ -1110,7 +1110,7 @@ static const char* kCLocale = "C";
 
     idx = main->uiPrefs->uiPrefsSubtitleFont->value();
     num = main->uiPrefs->uiPrefsSubtitleFont->children();
-    if ( idx < num )
+    if ( (int)idx < num )
     {
         const char* font = main->uiPrefs->uiPrefsSubtitleFont->child(idx)->label();
         CMedia::default_subtitle_font( font );
